requirements.txt
clipspy==1.0.4
tk